+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Don't throw an error if it's not a PR
 
+### Added
+
+-   Allow to specify ignored paths
+
 ## [1.1.0] - 2022-12-12
 
 ### Added